+++ encode.hpp
@@ -50,7 +50,7 @@ public:
 	    FILE *fp,
 	    const Music_info &track,
 	    int64_t total_samples,
-	    int64_t sample_rate,
+	    int32_t sample_rate,
 	    file_format=file_format::FLAC);
 
 	//! \throw Encode_error
